+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
+## v0.53.0 | 17 Apr 2023
+
+This update focuses on wind direction time series and adds the first example notebooks
+to `diive`. From now on, new example notebooks will be added regularly.
+
+## New features
+
+- **Wind direction offset correction**: Compare yearly wind direction histograms to
+  reference, detect offset in comparison to reference and correct wind directions
+  for offset per year (`pkgs.corrections.winddiroffset.WindDirOffset`)
+- **Wind direction aggregation**: Calculate mean etc. of wind direction in
+  degrees (`core.funcs.funcs.winddirection_agg_kanda`)
+
+## Notebooks
+
+- Added new notebook for **wind direction offset correction** in `notebooks/Corrections/WindDirectionOffset.ipynb`
+- Added new notebok for **reading ICOS BM files** in `notebooks/ReadFiles/Read_data_from_ICOS_BM_files.ipynb`
+
+## Changes
+
+- **Histogram analysis** now accepts pandas Series as input (`pkgs.analyses.histogram.Histogram`)
+
+## Additions
+
+- Added unittests for reading (some) filetypes
+
 ## v0.52.7 | 16 Mar 2023
 
 ## Additions